fix potential overflow address

+++ src/extensions/k12_engine.h
@@ -366,12 +366,12 @@ public:
                     auto accessAddress = msg.arg.pagefault.address;
                     auto pageAddress = msg.arg.pagefault.address & ~(page_size - 1);
 
-                    size_t offset = accessAddress - (uint64_t)_state;
+                    size_t offset = accessAddress - (size_t)_state;
                     unsigned int chunkIndex = offset / K12_chunkSize;
 
-                    auto startRange = (uint64_t)_state + (chunkIndex * K12_chunkSize);
+                    auto startRange = (size_t)_state + (chunkIndex * (size_t)K12_chunkSize);
                     // if there is only 1 page left (system memory page) then just need to cover to the last system page (cover full K12 chunk will go beyond the state size)
-                    size_t lenRange = std::min(paddedSize - (chunkIndex * K12_chunkSize), (size_t)K12_chunkSize);
+                    size_t lenRange = std::min(paddedSize - (chunkIndex * (size_t)K12_chunkSize), (size_t)K12_chunkSize);
                     // handle write-protect page fault
                     if (is_wp)
                     {
